[fix] 修改文件命名,把lk电机的特有api放在自己的头文件下
