+++ drivers/Kconfig
@@ -1,4 +1,4 @@
 rsource "./buzzer/Kconfig"
-rsource "./motor/Kconfig"
+rsource "./bldcm/Kconfig"
 rsource "./can_tx_manager/Kconfig"
 rsource "./can_rx_manager/Kconfig"
