+++ dts/bindings/bldcm/rp,lk-can-motor.yaml
