+++ samples/boards/dm_mc02/bldcm/CMakeLists.txt
@@ -1,9 +1,10 @@
 cmake_minimum_required(VERSION 3.20)
 
+set(Python3_EXECUTABLE "$ENV{WEST_PYTHON}" CACHE FILEPATH "Python 3 interpreter")
 if(NOT DEFINED BOARD)
 	set(BOARD damiao_mc02)
 endif()
 find_package(Zephyr REQUIRED HINTS $ENV{ZEPHYR_BASE})
-project(motor_sample)
+project(bldcm_sample)
 
 target_sources(app PRIVATE src/main.cpp)
