+++ samples/boards/dm_mc02/bldcm/Kconfig
