+++ samples/boards/dm_mc02/bldcm/prj.conf
@@ -2,7 +2,7 @@ CONFIG_CAN=y
 CONFIG_CPP=y
 CONFIG_CAN_RX_MANAGER=y
 CONFIG_CAN_TX_MANAGER=y
-CONFIG_MOTOR=y
+CONFIG_BLDCM=y
 CONFIG_ASSERT=y
 
 # DJI 电机 CAN 通常为 1Mbps；
@@ -13,8 +13,8 @@ CONFIG_CAN_DEFAULT_BITRATE=1000000
 # CONFIG_CAN_RX_MANAGER_MSGQ_WARN_EVERY_N_DROPS=1
 
 # DJI 心跳自动检测（shared system workqueue）
-CONFIG_MOTOR_HEARTBEAT_AUTOCHECK=y
-CONFIG_MOTOR_HEARTBEAT_POLL_PERIOD_MS=100
+CONFIG_BLDCM_HEARTBEAT_AUTOCHECK=y
+CONFIG_BLDCM_HEARTBEAT_POLL_PERIOD_MS=100
