+++ samples/boards/dm_mc02/bldcm/rm_typec.overlay
